Dashboard: add a view that shows tests in order of number of failures (#16)
**Context:**
 
This bug is about showing a simple view of the tests that flake the
most. We don't need to show all tests, but maybe it would be good to
show the top 100 tests and the number of times they had unexpected
results.

**Solution:**

- Created an app route /most-flaky for a view, which contains a table
that shows Test Path and Number of Unexpected Results
- Added a new function called `most_flaky` to retrieve the flakiest
tests from the database. This function calculates the total unexpected
count for each test and sorts them in descending order.
- Added a limit parameter as a query parameter of the request

- [x] This PR fixes #15 
- [x] `python3 -m intermittent_tracker.tests` doesn't report any error

### Test plan
- Successful test run:
`python3 -m intermittent_tracker.tests` doesn't report any error`
- Demo of a newly created view: